ExportToPDF function and visual done
