+++ client/src/components/WeeklyMenu/DishCard.jsx
@@ -1,207 +1,6 @@
 import PropTypes from 'prop-types';
 
-const DishCard = ({ item, getImageURL }) => {
-  const ingredientCategories = {
-    Mushrooms: ['Portobello Mushroom', 'Mushrooms'],
-    Herbs: [
-      'Chives',
-      'Cilantro',
-      'Basil',
-      'Thyme',
-      'Coriander',
-      'Rosemary',
-      'Basil Pesto',
-      'Parsley',
-      'Wasabi',
-    ],
-    Green: [
-      'Spinach',
-      'Celery',
-      'Zucchini',
-      'Lettuce',
-      'Asparagus',
-      'Cabbage',
-      'Peas',
-      'Romaine Lettuce',
-    ],
-    Cheese: [
-      'Mozzarella',
-      'Ricotta Cheese',
-      'Mozzarella Cheese',
-      'Vegan Cheese',
-      'Parmesan Cheese',
-      'Parmesan',
-      'Cheese',
-      'Cheddar Cheese',
-      'Vegan Cheese Sauce',
-      'Blue Cheese',
-    ],
-    Garlic: ['Garlic', 'Garlic Powder'],
-    Olives: ['Olive Oil', 'Olives'],
-    Rice: ['Quinoa', 'Arborio Rice', 'Rice'],
-    Avocados: ['Avocado', 'Guacamole'],
-    Nori: ['Nori'],
-    Flour: ['Flour', 'Almond Flour', 'Gluten-Free Oats', 'Gluten-Free Flour'],
-    Chocolate: ['Cocoa Powder', 'Chocolate Chips'],
-    Milk: ['Almond Milk', 'Cream', 'Butter', 'Heavy Cream', 'Milk'],
-    Others: [
-      'Sugar',
-      'Baking Powder',
-      'Soy Sauce',
-      'Salt',
-      'Curry Powder',
-      'Taco Seasoning',
-      'Cinnamon',
-      'Nutritional Yeast',
-      'Turmeric',
-      'Cumin',
-    ],
-    Dressing: [
-      'Hummus',
-      'Caesar Dressing',
-      'Mayo',
-      'Gluten-Free Soy Sauce',
-      'Balsamic Vinegar',
-      'Maple Syrup',
-      'Tamari Sauce',
-      'Vegan Caesar Dressing',
-    ],
-    Seafood: [
-      'Clams',
-      'Tuna',
-      'Shrimp',
-      'Salmon',
-      'White Fish',
-      'Cod',
-      'Lobster',
-      'Mussels',
-    ],
-    Potatoes: ['Potatoes', 'Mashed Potatoes'],
-    Onions: ['Onion', 'Shallots'],
-    Meat: [
-      'Chicken Breast',
-      'Beef',
-      'Ground Chicken',
-      'Pepperoni',
-      'Ground Beef',
-      'Ground Turkey',
-      'Turkey',
-      'Chicken',
-      'Chicken Thighs',
-      'Bacon',
-      'Beef Strips',
-    ],
-    Tomatoes: [
-      'Tomato',
-      'Tomato Sauce',
-      'Tomato Sauce',
-      'Tomatoes',
-      'Salsa',
-      'Cherry Tomatoes',
-    ],
-    Cucumber: ['Cucumber', 'Pickles'],
-    Broth: ['Broth', 'Vegetable Broth'],
-    Peppers: ['Bell Pepper', 'Bell Peppers'],
-    Chilli: ['Chili Powder', 'Paprika'],
-    Bread: [
-      'Whole Wheat Bun',
-      'Breadcrumbs',
-      'Gluten-Free Pizza Dough',
-      'Croutons',
-      'Whole Wheat Wrap',
-      'Gluten-Free Pizza Crust',
-      'Pizza Dough',
-      'Burger Bun',
-      'Vegan Bun',
-      'Brioche Bun',
-      'Bread',
-      'Pastry',
-      'Gluten-Free Breadcrumbs',
-    ],
-    Carrots: ['Carrots'],
-    Noodles: [
-      'Gluten-Free Pasta',
-      'Fettuccine',
-      'Linguine',
-      'Lasagna Noodles',
-      'Spaghetti',
-      'Lasagna Sheets',
-      'Spaghetti Squash',
-      'Pasta',
-      'Rice Noodles',
-    ],
-    Citrus: ['Lemon Juice', 'Lime Juice', 'Lemon', 'Lemon Dressing', 'Lime'],
-    Tacos: ['Corn Tortillas', 'Taco Shells', 'Flour Tortilla'],
-    Oil: ['Oil'],
-    wine: ['White Wine'],
-    apples: ['Apples'],
-    Nuts: ['Tahini', 'Sesame Seeds', 'Pine Nuts', 'Cashews'],
-    Bean: ['Tofu', 'Chickpeas', 'Black Beans', 'Falafel', 'Lentils'],
-    Peanuts: ['Peanuts'],
-    Egg: ['Egg', 'Eggs'],
-    Coconut: [
-      'Coconut Oil',
-      'Coconut Milk',
-      'Coconut Aminos',
-      'Coconut Flakes',
-    ],
-    Eggplants: ['Eggplant'],
-    'Sweet Potato': ['Sweet Potatoes'],
-    Ginger: ['Ginger'],
-    Cauliflower: ['Cauliflower Rice', 'Broccoli', 'Cauliflower'],
-  };
-
-  const ingredientEmojis = {
-    Mushrooms: '🍄',
-    Green: '🥬',
-    Herbs: '🌿',
-    Cheese: '🧀',
-    Garlic: '🧄',
-    Olives: '🫒',
-    Rice: '🍚',
-    Avocados: '🥑',
-    Nori: '🍣',
-    Flour: '🌾',
-    Chocolate: '🍫',
-    Milk: '🥛',
-    Others: '🧂',
-    Dressing: '🫙',
-    Seafood: '🐟',
-    Potatoes: '🥔',
-    Onions: '🧅',
-    Meat: '🥩',
-    Tomatoes: '🍅',
-    Cucumber: '🥒',
-    Broth: '🍲',
-    Peppers: '🫑',
-    Chilli: '🌶️',
-    Bread: '🍞',
-    Carrots: '🥕',
-    Noodles: '🍝',
-    Citrus: '🍋',
-    Tacos: '🌮',
-    Oil: '🛢️',
-    wine: '🍷',
-    apples: '🍎',
-    Nuts: '🌰',
-    Bean: '🫘',
-    Peanuts: '🥜',
-    Egg: '🥚',
-    Coconut: '🥥',
-    Eggplants: '🍆',
-    'Sweet Potato': '🍠',
-    Ginger: '🫚',
-    Cauliflower: '🥦',
-  };
-
-  const getIngredientEmoji = (ingredient) => {
-    for (const [category, ingredients] of Object.entries(ingredientCategories))
-      if (ingredients.includes(ingredient)) {
-        return ingredientEmojis[category] || '😊';
-      }
-    return '🤭';
-  };
-
+const DishCard = ({ item, getImageURL, getIngredientEmoji }) => {
   const formatSelectedDate = (dateString) => {
     const date = new Date(dateString);
     const options = { weekday: 'long', day: '2-digit' };
